Tests on Combo all good

+++ food-app-backend/app/Http/Controllers/Api/Combos/ComboController.php
@@ -17,9 +17,7 @@ class ComboController extends Controller
 
     public function __construct(private readonly ComboService $comboService)
     {
-        $this->authorizeResource(Combo::class, 'combo', [
-            'except' => ['index', 'show'],
-        ]);
+        //
     }
 
     public function index(Request $request)
@@ -30,8 +28,20 @@ class ComboController extends Controller
         return $this->success(ComboResource::collection($combos));
     }
 
+    public function getRestaurantCombos(string $id, Request $request)
+    {
+        $withRelations = $request->boolean('include_relations', true);
+        $combos = Combo::where('restaurant_id', $id)
+            ->when($withRelations, fn($query) => $query->with(['restaurant', 'groups.items.dish']))
+            ->get();
+
+        return $this->success(ComboResource::collection($combos));
+    }
+
     public function store(StoreComboRequest $request)
     {
+        $this->authorize('create', Combo::class);
+        
         $combo = $this->comboService->create($request->validated());
 
         return $this->success(new ComboResource($combo), 'Combo created successfully', 201);
@@ -46,6 +56,8 @@ class ComboController extends Controller
 
     public function update(UpdateComboRequest $request, Combo $combo)
     {
+        $this->authorize('update', $combo);
+        
         $combo = $this->comboService->update($combo, $request->validated());
 
         return $this->success(new ComboResource($combo), 'Combo updated successfully');
@@ -53,6 +65,8 @@ class ComboController extends Controller
 
     public function destroy(Combo $combo)
     {
+        $this->authorize('delete', $combo);
+        
         $this->comboService->delete($combo);
 
         return $this->success(null, 'Combo deleted successfully');
